+++ src/grammar/next/grammar.spec.ts
@@ -1,7 +1,7 @@
-import { getAstFactory } from "./ast";
-import { loadCases } from "../utils/loadCases";
-import { getParser } from "./grammar";
-import { SrcInfo, isSrcInfo } from "./src-info";
+import { getAstFactory } from "../ast";
+import { loadCases } from "../../utils/loadCases";
+import { getParser } from "../grammar";
+import { SrcInfo, isSrcInfo } from "../src-info";
 
 expect.addSnapshotSerializer({
     test: (src) => isSrcInfo(src),
@@ -9,18 +9,18 @@ expect.addSnapshotSerializer({
 });
 
 describe("grammar", () => {
-    for (const r of loadCases(__dirname + "/test/")) {
+    for (const r of loadCases(__dirname + "/../test/")) {
         it("should parse " + r.name, () => {
             const ast = getAstFactory();
-            const { parse } = getParser(ast);
+            const { parse } = getParser(ast, 'new');
             expect(parse(r.code, "<unknown>", "user")).toMatchSnapshot();
         });
     }
 
-    for (const r of loadCases(__dirname + "/test-failed/")) {
+    for (const r of loadCases(__dirname + "/../test-failed/")) {
         it("should fail " + r.name, () => {
             const ast = getAstFactory();
-            const { parse } = getParser(ast);
+            const { parse } = getParser(ast, 'new');
             expect(() =>
                 parse(r.code, "<unknown>", "user"),
             ).toThrowErrorMatchingSnapshot();